+++ src/test/java/io/github/mmalykhin/hmsproxy/frontend/FrontendProcessorFactoryTest.java
@@ -67,6 +67,7 @@ public class FrontendProcessorFactoryTest {
             "catalog1", "c1", "file:///c1", false, ProxyConfig.CatalogAccessMode.READ_WRITE, java.util.List.of(),
             null, null, Map.of("hive.metastore.uris", "thrift://one"))))
         .compatibility(new ProxyConfig.CompatibilityConfig(ProxyConfig.FrontendProfile.APACHE_3_1_3, null, null, false))
+        .syntheticReadLockStore(ProxyConfig.SyntheticReadLockStoreConfig.inMemory())
         .build();
   }
 
@@ -80,6 +81,7 @@ public class FrontendProcessorFactoryTest {
             "catalog1", "c1", "file:///c1", false, ProxyConfig.CatalogAccessMode.READ_WRITE, java.util.List.of(),
             null, null, Map.of("hive.metastore.uris", "thrift://one"))))
         .compatibility(new ProxyConfig.CompatibilityConfig(frontendProfile, jar.toString(), null, false))
+        .syntheticReadLockStore(ProxyConfig.SyntheticReadLockStoreConfig.inMemory())
         .build();
   }